add all models

+++ src/main/java/com/piseth/java/school/phoneshopenight/service/BrandService.java
@@ -9,8 +9,8 @@ import com.piseth.java.school.phoneshopenight.entity.Brand;
 
 public interface BrandService {
 	Brand create(Brand brand);
-	Brand getById(Integer id);
-	Brand update(Integer id, Brand brandUpdate);
+	Brand getById(Long id);
+	Brand update(Long id, Brand brandUpdate);
 	List<Brand> getBrands(String name);
 	//List<Brand> getBrands(Map<String, String> params);
